Corrected bug pointed out by Phil Seeger / Kristian Nielsen, from their correspondance:
PASeeger@aol.com writes:

> Kristian,
>
> Please check this statement in single_crystal.comp:
>           mos_e2_x = -mos_c - mos_l2;
>
> I believe it should be plus mos_c, not minus.

Yes, going through the calculation it is pretty obvious that this is a
bug. The code is computing orthogonal eigenvectors for a 2-by-2
matrix. Using your correction, the computed vectors are indeed both
eigenvectors of the matrix and orthogonal to each other. Using the
original code they are neither :-(. Well spotted, Phil (and shame on me).

Kim/Peter, you should change the code as suggested by Phil. In the
McStas 1.4.2 version of the file, it is line 288 that must be changed to

        mos_e2_x = mos_c - mos_l2;

> I was surprised to find that the eigenvectors for the anisotropic matrix were
> not orthogonal!  This sign change fixes that, and though my comfort level in
> reciprocal space is approximately nil, it appears to me that the principal
> axes of the anisotropic Gaussian now track with a rotated unit cell.

Great!

 - Kristian.