Platform independence:
On Solaris, the printf function seems to dislike null-pointers. This means that when HOSTNAME is not a defined environment variable, h8_test fails with a runtime error (SIGMEM).

Thought about using the 'hostname' command instead, but doing this properly requires handling of subprocess stdio -> tedious.

Hence, the string "(undefined)" is used in cases where HOSTNAME is undefined.

(BTW, it seems that indeed this variable is undefined on many distros.)