First try debian build system.
In case of a debian build host (/etc/debian_version is there), the debcreate
script is run, resulting in .deb and .rpm packages (the latter using alien -
clearly needs testing).

The deb package is likely (clearly!!) not fully debian compliant but seems
to work ok. It includes pgplot and PGPERL - the linux alternative to the
new Win32 .exe.