Including patches from Jan Saroun (NPI) - explanation from JS email:
"this is an update of PowderN, which solves the problem with target focusing. The original component didn't work well when rotated like

ROTATED (0, omega, chi) RELATIVE Sample_axis

it focused neutrons out of the diffractometer plane. The new version permits to set target_index and focus towards the detector. Also incoherent scattering can be focused within a selected angle in this direction. Statistics is then much improved in cases like an engineering diffractometer (sample in various orientations, limited detector coverage).  I guess it could be easily merged with current PowderN if you find it useful. There are new parameters added, but behavior should not change if default values are used.
"