Work in progress on updating the Union components to avoid function pointers, which are used in many places. This version avoids these function pointers in the geometrical part, and have functions ready for the physics part.
Instead of carrying a function pointer to the required function,
each relevant struct has an enum that specifies which function
should be called in its case. A middle function will manage this
and call the right function using a case statement.

ifdef defines are used to avoid compiling scattering processes that
are not currently used.