Update of Union components, mainly adding the new absorption loggers.
The absorption loggers can record absorption events in the same way
that regular loggers can record scattering events. They can also be
affected by conditional components in the same way as regular logger
components.

A new example was added which demonstrate and test the new components.

Some issues with the mesh component was spotted during this work, it
seems to work well under 2.5, but breaks in 2.6 or newer. This is not
yet understood.