Initial commit of NCrystal Union process based on the NCrystal_sample component provided by the NCrystal development team.
The NCrystal_sample component includes absorption and a geometry
representation, but this is not possible for a Union process as these
tasks are performed by other parts of the Union system.

This version supports NCrystal version 1.0.0, yet a new 2.0.0 version
will be released soon. Updates that enables support for 2.0.0 will
be made soon.