Update of the Union components that now work on cogen 3. Some testing still to be done, but by commiting to this branch much of this testing is automated and will be logged.
Added new absorption logger category of components that log absorption.

Added initialization component which is currently needed as a replacement
for global variables. All subsequent union components need to know the
name of this component, but a default of init is used in every case, so
if that name is used for the initialization component, nothing else needs
to be updated. All examples had this added.

Bugs in memory management were discovered and removed.

the abs_logger_1D_space_tof_to_lambda needs POS_A_COMPONENT which seems
to be broken in the current version of the cogen3.