Manual merge of branch without polymorphism into McStas 3.0 branch.
Major change is that function points in trace are being removed,
currently for intersection call, cross section calculation and
scattering function. Each replaced by a function with a switch
around the necessary functions. The ones pertaining to processes
are activated / deactivated based on whether they are in the
instrument or not by flags in the preprocessor. Since these flags
need to be set before the function is defined but after all
processes are defined, a stop component is introduced which should
be after all Union components in the instrument file.

All examples are updated to include such a stop component.

Loggers and Abs_loggers still use function pointers and need
further work to get rid of these.