Fixed absorption logger + p_interact bug in Union_master.
When setting p_interact, negative absorption weights could be less
than zero due to wrong approach in the weight calculation. Now the
p_interact manipulation is not considered absorption, but a change
in the weight before absorption is applied.

Was verified with beers law and against same code without p_interact.