[Refactor] CMake should pull-in MCPL and NCrystal
Both MCPL and NCrystal are built along side McStas & McXtrace. These
should only be pulled from a remote repository and built if they are not
present locally.
This commit adds CMake modules to find and fetch if not present any
general CMake project, plus to do so for the MCPL and NCrystal projects.

The CMakeList files are updated to provide new option flags to build
either. The old behvior of requiring externally-populated 'thirdparty'
directories is maintained by default.