Optimization of code generation in cases where components have no trace and is not referred to with RELATIVE. These components can be skipped in the generated traceloop in most respects, and can be significantly faster because coordinate transformations are skipped and debug state is not written.
The setpos functions are still added, but their content reffers to the
last non skipped component.

There may be issues with jump, which can be rectified by not calling
a component useless if a jump is detected.

Significant speed up especially for the Union components, but only
if the geometry components are not used for RELATIVE by a usefull
component.