New version of traceless component optimization.
Now removes component transforms if the following is true:
  Has no trace
  Has no extend
  Does not jump somewhere
  Is not jumpped to

This can skip nearly all Union components and arms. The code is
quite straight forward as there is no graph analysis anymore.
Thus there is no danger of using RELATIVE in a way that results in
large changes in performance which in the first optimization
attempt would loose the ability to prune a sequence of traceless
components.