Update MCPL_output.comp filename
The MCPL_output documentation string indicated that a unique filename would be used if no filename was provided. This was not true, and instead the filename `[output_directory]/.mcpl` would be used. Multiple MCPL_output components would use the same filename if not given a filename value, and either wrote to the same file or overwrote each other depending on implementation details in MCPL. Furthermore, the produced filename is hidden on UNIX-like systems, so a user may never find their output file.

The component has been changed to insert the (unique) component name into the filename string if another value is not supplied. This should produce MCPL output files like `[output_directory]/[component_name].mcpl` which a) are unique and b) are not hidden.